Move application configuration to settings class
